add :init to forOptimisation switch ... so that productreturns table can be 'truncate'd prior to multiple calls to forOptimisation ... speed up by parallelising forOptimisation

+++ ConsoleApplication1/ConsoleApplication1/ConsoleApplication1.cpp
@@ -93,7 +93,7 @@ int _tmain(int argc, WCHAR* argv[])
 		argWords["stickySmile"]             = "";
 		argWords["bump"]                    = "bumpType:startBump:stepSize:numBumps eg delta|vega|rho|credit|corr~name~name:-0.05:0.05:3 >";
 		argWords["bumpVolPoint"]            = "tenor:strike:bumpAmount(decimal) eg 1.0:0.6:0.01 ";
-		argWords["forOptimisation"]         = "";
+		argWords["forOptimisation"]         = "init|any";
 		argWords["duration"]                = "<number ~ number, or just number(min)>";
 		argWords["volatility"]              = "<number ~ number, or just number(min) PERCENT>";
 		argWords["arithReturn"]             = "<number ~ number, or just number(min) PERCENT>";
@@ -126,7 +126,7 @@ int _tmain(int argc, WCHAR* argv[])
 		int              debugLevel(0),corrUidx(0), corrOtherUidx(0), corrOtherIndex(0), doUseMyEqEqCorr(-1), doUseMyEqFxCorr(-1);
 		bool             doTesting(false),doFinalAssetReturn(false), requesterForceIterations(false), doDebug(false), getMarketData(false), notStale(false), hasISIN(false), hasInventory(false), notIllustrative(false), onlyTheseUls(false), forceEqFxCorr(false), forceEqEqCorr(false);
 		bool             doUseThisVolShift(false), doUseThisBarrierBend(false), doUseThisOIS(false), doUseThisPrice(false), showMatured(false), doBumps(false), doDeltas(false), doPriips(false), ovveridePriipsStartDate(false), doUKSPA(false), doAnyIdTable(false);
-		bool             doRescale(false), doRescaleSpots(false), doBarrierBendAmort(true) /* lets try it */, doStickySmile(false), useProductFundingFractionFactor(false), forOptimisation(false), silent(false), updateProduct(false),verbose(false), doIncomeProducts(false), doCapitalProducts(false), solveFor(false), solveForCommit(false);
+		bool             doRescale(false), doRescaleSpots(false), doBarrierBendAmort(true) /* lets try it */, doStickySmile(false), useProductFundingFractionFactor(false), forOptimisation(false), initOptimisation(false), silent(false), updateProduct(false),verbose(false), doIncomeProducts(false), doCapitalProducts(false), solveFor(false), solveForCommit(false);
 		bool             bsPricer(false),forceLocalVol(false),localVol(true), stochasticDrift(false), ignoreBenchmark(false), done, forceFullPriceRecord(false), fullyProtected, firstTime, forceUlLevels(false),corrsAreEqEq(true);
 		bool             doFvEndDate(false),updateCashflows(true),ajaxCalling(false),slidingTheta(false),cmdLineBarrierBend(false);
 		
@@ -541,6 +541,7 @@ int _tmain(int argc, WCHAR* argv[])
 			if (sscanf(thisArg, "Issuer:%s", lineBuffer))                 { issuerPartName          = lineBuffer; }
 			if (sscanf(thisArg, "fundingFractionFactor:%s",   lineBuffer)){ fundingFractionFactor	= atof(lineBuffer);	}
 			if (sscanf(thisArg, "forceFundingFraction:%s",    lineBuffer)){ forceFundingFraction	= lineBuffer; }		
+			else if (sscanf(thisArg, "forOptimisation:%s", lineBuffer)) { forOptimisation    = true; if (strcmp("init",lineBuffer) == 0) { initOptimisation = true; }  }
 			else if (sscanf(thisArg, "spotsDate:%s",          lineBuffer)){ strcpy(spotsDate, lineBuffer); }
 			else if (doFvEndDate || sscanf(thisArg, "endDate:%s",       endDate)){ strcpy(endDate,        endDate); }
 			else if (doFvEndDate || sscanf(thisArg, "arcVolDate:%s",    endDate)){ strcpy(arcVolDate,     endDate); sprintf(arcVolDateString,    "%s%s%s", " and LastDataDate='", arcVolDate,     "' "); }
@@ -730,10 +731,12 @@ int _tmain(int argc, WCHAR* argv[])
 			mydb.prepare((SQLCHAR *)lineBuffer, 1);
 
 			// clear productreturns and simulatedunderlyings tables
-			sprintf(lineBuffer, "%s", "delete from productreturns");
-			mydb.prepare((SQLCHAR *)lineBuffer, 1);
-			sprintf(lineBuffer, "%s", "delete from simulatedunderlyings");
-			mydb.prepare((SQLCHAR *)lineBuffer, 1);
+			if (initOptimisation) {
+				sprintf(lineBuffer, "%s", "delete from productreturns");
+				mydb.prepare((SQLCHAR *)lineBuffer, 1);
+				sprintf(lineBuffer, "%s", "delete from simulatedunderlyings");
+				mydb.prepare((SQLCHAR *)lineBuffer, 1);
+			}
 		}
